requirements.txt
discord.py
flask
pyotp
